+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:sampleproject/my_app.dart';
+import 'package:sampleproject/src/app.dart';
 
 void main() {
   runApp(const MyApp());
