+++ lib/src/app.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:sampleproject/src/home_page.dart';
+import 'package:sampleproject/src/pages/home_page.dart';
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       title: 'Flutter Demo',
       theme: ThemeData(
