--- lib/src/home_page.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children:[
-          Image.network("https://images.unsplash.com/photo-1611915387288-fd8d2f5f928b?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=1080&fit=max"),
-          Container(
-        decoration:  BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topLeft,
-            end: Alignment.bottomRight,
-            colors: [
-              Colors.transparent,
-
-              Color(0xffff0e35).withOpacity(0.8),
-           //   Color(0xff000000),
-            ]
-          )
-        ),
-          ),
-        ]
-      ),
-    );
-  }
-}
